moved the settings from index.js to /settings/index.js
