+++ components/Navigation/index.js
@@ -1,0 +1,10 @@
+import Link from "next/link";
+
+export default function NavBar() {
+  return (
+    <>
+      <Link href="/barcodeScanner">scan</Link>&nbsp;
+      <Link href="/settings">settings</Link>
+    </>
+  );
+}
