+++ next.config.js
@@ -15,10 +15,10 @@ const nextConfig = {
   },
   images: {
     remotePatterns: [
-      // {
-      //   protocol: "https",
-      //   hostname: "images.unsplash.com",
-      // },
+      {
+        protocol: "https",
+        hostname: "images.openfoodfacts.org",
+      },
     ],
   },
 };
