+++ pages/_app.js
@@ -1,14 +1,19 @@
+import NavBar from "@/components/Navigation";
 import GlobalStyle from "@/styles";
 import Head from "next/head";
+import { SWRConfig } from "swr";
 
 export default function App({ Component, pageProps }) {
   return (
     <>
-      <GlobalStyle />
-      <Head>
-        <title>Capstone Project</title>
-      </Head>
-      <Component {...pageProps} />
+      <SWRConfig>
+        <GlobalStyle />
+        <Head>
+          <title>Capstone Project</title>
+        </Head>
+        <Component {...pageProps} />
+        <NavBar></NavBar>
+      </SWRConfig>
     </>
   );
 }
