+++ pages/index.js
@@ -1,47 +1,16 @@
 import { useState } from "react";
 import allergens from "../allergens.json";
 import additives from "../additives.json";
-
 import { atomWithStorage, createJSONStorage } from "jotai/utils";
 import Card from "@/components/Card";
-
-const initialAdditives = atomWithStorage("additives", [], {
-  ...createJSONStorage(() => localStorage),
-  delayInit: true,
-});
-
-const initialAllergens = atomWithStorage("allergens", [], {
-  ...createJSONStorage(() => localStorage),
-  delayInit: true,
-});
+import { useRouter } from "next/router";
 
 export default function HomePage() {
-  const [showAdditives, setShowAdditives] = useState(true);
-
   return (
     <>
-      <button
-        style={{ fontWeight: showAdditives ? "bold" : "normal" }}
-        onClick={() => {
-          setShowAdditives(true);
-        }}
-      >
-        Additives
-      </button>
-      <button
-        style={{ fontWeight: !showAdditives ? "bold" : "normal" }}
-        onClick={() => {
-          setShowAdditives(false);
-        }}
-      >
-        Allergens
-      </button>
+      <h1>Welcome to Eatable</h1>
 
-      {showAdditives ? (
-        <Card initialItems={initialAdditives} items={additives} />
-      ) : (
-        <Card initialItems={initialAllergens} items={allergens} />
-      )}
+      <p>first go to the settings page </p>
     </>
   );
 }
